frontend: use smaller webm for the unlock gesture animation
The animated gif that is shown on the unlock view is roughly about
10MB and loaded as an image, this is ok, but in webdev the image
is loaded as base64 string in JavaScript and part of the main
bundle. This slows down frontend development especially if dev
tools are open.

With WEBM the video is now small enough that there is no noticable
difference anymore when loading the unlock view in webdev.

- allow importing .webm so TypeScript doesn't complain
- fixed autoplay when rerendering the view, i.e. due to re-plug the
  BitBox as Preact 8 does cache the DOM elements and the engine
  pauses the video once it is removed from the DOM