android: request USB permission if not automatically granted
There are these scenarios to cover:

- App is closed, user inserts device:

In this case, the user is prompted with all apps that register intent
filters for the device, including the BitBoxApp.

a) If the BitBoxApp is chosen, the app is launched with the
`ACTION_USB_DEVICE_ATTACHED` intent. Previously we would simply update
the device list in `onResume()` without checking the intent, which
included this case. Now we explicitly check the intent in
`onResume()` to avoid endlessly asking the user for permission (see below).

b) If the app is not chosen, permission is denied. If the user then
opens the app (without USB permission granted automatically like
above), previously the app would crash because of it. In this commit,
we request permission in `onCreate()` (by calling
`updateDevice()`). We do it newly in `onCreate()` instead of
`onResume()` because if permission is denied, `onResume()` is called
again, asking the user again for permission endlessly.

- App is opened, user inserts device:

App chooser is shown to the user like above. If they choose the same
BitBoxApp, `onResume()` is called with the intent, same as above.

If they reject, the app remains open without any further action.

- User opens app while the device is already connected

This only happens if the user rejects to open the BitBoxApp after
being prompted.

In this case, `onCreate()` calls `updateDevice()` which asks for USB permission.