config: change default fiat currency to usd
Currently the fiat currency default is Swiss Franc (CHF), but many
people don't know CHF or this default is not very useful for most.

Changed default to US Dollar (USD) as it is the most widely used
pair with Bitcoin and ohter crypto currencies. Kept Euro and CHF
as enabled secondary fiat currencies.