frontend: remove animations
Since updating to Qt 5.15, animations are very laggy on
Windows (opening dialogs).

It could be fixed by enabling GPU accelerated rendering, but that led
do various rendering artefacts and even crashes.

This is a compromise: by removing animations, the app is not laggy on
Windows even when only using software rendering.