frontend: refactor input component
Changed input component to use forwardRef instead of getRef and
updated input tests to use @testing-library/react.

Current issue:

The input component had a getRef callback prop to pass the ref
of the input element. This is used for fixing the focus in
transaction-notes, manage-accounts, BitBox01 (password, unlock,
create-seed) and checking validation in BitBox02 (setdevicename).

AutoFocus never properly worked in Preact 8 and also the other
uses where hacky and ideally should be done by the component
itself.
Comparing passwords in BitBox01 should be done through state.

Using forwardRef hook allows to pass the components inner input
element, so that the parent can pick it up with useRef hook.
Unfortunatelly there is no good way to write tests for this.
Also hooks (useRef) can only be used in parents that are using
hooks as well and cannot be used in parents that still use
React.Component class.
Also forwardRef is an anti-pattern (apparently) and everything
could be done by the component itself.

Ant design uses 2 callbacks for form validation: onFinish and
onFinishFailed with errorInfo.

Proposal:
- only use autoFocus prop, remove hacks that try to focus in the
  getRef callback.
- use state and move validation to onInput