frontend: mnemonicpassphrase component only needs passenabled prop
Currently the whole deviceInfo object is passed into the mnemonic
passphrase component, but only mnemonicPassphraseEnabled (boolean)
is used from this object and the rest is not needed.

Changed to accept mnemonicPassphraseEnabled (boolean) via props
and in state, that way react doesn't have to deep diff the object
and improves readability.