frontend: improve information about using the passphrase feature
It is a regular occurrence that a user asks for support on Telegram
about funds or transactions missing, because they had used a
passphrase and did not understand what it meant. Often the issue
pops up when restoring a wallet (passphrase needs to be activated
and entered again), and also some users simply deactivate the
passphrase and are confused why the funds are gone.

Users considering enabling passphrase should be well informed and
know about the advantages and risks.

Added what, why, how, backup infos explanations on different pages
to inform and improve using the passphrase feature.

Made the success step blocking with a prompt to force replug the
BitBox. This is so that the changed settings take effect.