frontend: password input component autofocus on mount
Previous to this change the BitBox01 unlock screen created a
reference to the password input and tried to focus it. AutoFocus
never worked properly with Preact 8 and becuase of that there is
logic in the code that tries to focus the input element.

This blocked createRef migration which changed the ref to using
hook-style refs with createRef. Before password input component
passed in a callback to get the ref of the input. With the
createRef migration this changed and the input element was not
found anymore and error when trying to focus.

Changed password input components to check for autoFocus prop
and trying to focus the element on mount. Tested in Qt app as it
has slightly different behavior than in the browser with webdev.