frontend: remove old logic to clear password repeat input
In BitBox01 the device password is entered in the BitBoxApp, for
setting the initial device or backup password we have a component
that renders 2 password input elements and checks if they have
the same value.

In BitBox01 initialize and create-new-seed there was custom logic
to clear the password input (child component), but this broke
with the migration to createRef (road to React).

This logic is not needed, as each time the view renders the
passwordrepeat component it will mount new and set its internal
state to empty password.

Removed the logic to clear passwordrepeatinput.