frontend: fix routing on unplug or app restart
Changed setDevcies to setStateWithDeviceList to make it clear that
this is the only way devices should be set in the state, also this
conditionally calles maybeRoute if the function does not route
to manage-device.

Currently if there is a new device detected the routing logic
does assume this has to be unlocked, paired etc. and routes
to device settings. One edge case where this is wrong is when
the device is unlocked and the app is restarted, this can be
detected by checking if there are accounts on the newly detected
device.

Changed routing logic to: if the device is already unlocked route
to account-summary else to device settings.