frontend: control select component
React has a defaultValue property on select dropdowns to set the
default value, but this is intendet to be used with uncontrolled
components. To control a select element React uses value.
Using both defaultValue and value triggers a warning: Select
elements must be either controlled or uncontrolled (specify
either the value prop, or the defaultValue prop, but not both).

AOPP, buy/choose-account and manage-accounts/select-coin all had
a defaultValue and a value props. As far as I can see they could
be all uncontrolled as there is no other form field or other
data that could change the select.
Decided to go with controlled component using value and removing
defaultValue, because I stlightly prefer controlled, i.e. if we
ever add another field that may change the value in the future.