frontend: fix electrum ui initialization
There was a regression that the configured nodes only showed
TLS, but not what the actual address/domain is.

The problem was that the default state of the servers was
initialized based on props in the constructor. Using setState in
the constructor worked in Preact, but fails in React without an
error or warning.

Changed to intialize the state late in componentDidMount callback
so it shows now correct server settings.