frontend: show selected backup in a sentence in restore-from-sdcard
In the last release the selected backup has been added on top of
the restore step where to set the new password.

The backup name, date and id was shown without context.

This commit adds a sentence that mentiones that this backup is
selected and will be used to restore the backup from.