frontend: add usesync hook
With the subscribe HOC it made an initial request by default
and subscribed to future changes. This is currently not easy
to do with useLoad and useSubscribe, therefore added a combined
hook that does both.

The useSync hook takes a load promise and a subscription
function, returns undefined while loading and re-renders on every
update.