frontend: show missing error from backup component during setup
Prior to moving the setup completelly to the view component, some
of the steps were still mounted and could catch errors even if
that was not the active step.
By conditionally rendering different views this behavior changed
and the backuplist unmounted after continuing to set-password.

The issue is that if a wrong password is entered to confirm
setting a new password, the backup component cannot catch that
there was an error.

Reproduce:
1) restore from sd-card
2) select a backup and continue
3) repeat password with a different one
4) expect an error

Current fix:
Keep restore view with the backup component mounted even after
proceeding, this is how the component could listen and catch the
error before.