frontend: disable chart filter buttons if chartdatamissing
The chart filter buttons to change the chart to display
week, month, year and all is currently only checking
if the chart contains any data and the total is not 0.

But if the chart is updating the buttons are already active
even though the chart has not loaded and is displaying
updating data.

Changed to check for chartDataMissing and keep the buttons
disabled until everything is up to date.