backend/eth/rpcclient: remove unused RPClient
As the comment suggests, we kept it thinking that we might need it
again.

The reason for this was only used official RPC calls where possible,
so we could use any ETH RPC node if needed. This complicates the code
considerably, and it's also not viable, since we need a list of
transactions anyway, which is not part of the official RPC interface.