backend/conversions: BTC formatting bug fixed
The app offers to the user a list of currencies for coin value
conversion: most of the currencies are fiat and should be formatted
with two fractional digits. The list includes BTC, which should be
formatted with 8 fractional digits, but was instead handled like the
other currencies. This commit fix this, formattingv BTC converted
values with 8 digits.