backend/frontend: refactor coin/currency conversion and formatting
This commit refactor the code dedicated to currency and coin
conversions and formatting.

The formatting logic of coin and currencies amounts has been
refactored to be handled in a small number of well defined functions,
as a preparatory step for a later inclusion of l10n handling.

Formatting and conversion functions updated to exploit Big.Rat precision