backend/devices/keystore: don't panic if previous output missing
Not being able to reference the previous output of an input is a bug,
but at the moment one can run into it when making really fast
consecutive transactions: if a new tx is made before the previous one
is broadcast and received from the Electrum backend, it might reuse
UTXOs, which become spent the moment the previous tx is processed.

This commit simply removes the panic and turns it into an error that
the user sees in the frontend.