frontend/send: disable fiat selection for send view
The use of fiat selection in send view was affecting only the balance,
but not the other parts of the page (i.e. the fiat amount to be transfered).
Since it is not easy to decide which should be the correct behavior if
the user change the fiat while compiling the form, we thought it was
better to disable the feature in this view.