frontend/bitbox02: abort SetPassword when the user aborts
Before, the error was always assumed to be the password mismatch
error, and the SetPassword workflow was invoked again immediately.

Now we go back to the initial screen if the user aborts.

The BitBox02 only returns the user abot error from this API endpoint
from v9.13.0, before the two types of errors can not be distinguished.