frontend: enable sats unit option for bitcoin accounts
The main unit displayed in the app for bitcoin accounts is Bitcoin
(BTC), but sometimes it can be easier to think in terms of Satoshi
(sat), especially for lower amounts.

This allows the user to choose the preferred unit using a
switch in the expert settings section of the frontend.